Binary .py files indentation solution and Fix declarations in 'for' and 'uint32_t' in argmax function of binary libraries

+++ embedia/libraries/binary/embedia.h
@@ -353,7 +353,7 @@ void avg_pooling2d_layer(pooling2d_layer_t pool, data3d_t input, data3d_t* outpu
  * Returns:
  *  search result - index of the maximum value
  */
-uint16_t argmax(data1d_t data);
+uint32_t argmax(data1d_t data);
 
 
 /***************************************************************************************************************************/
